fix : the language change in the layout.tsx file
