--- src/app/loading.tsx
@@ -1,9 +1,0 @@
-const Loading = () => {
-  return (
-    <div className="flex justify-center items-center h-screen ">
-      <div className="w-12 h-12 border-4 border-blue-500 border-t-transparent rounded-full animate-spin" />
-    </div>
-  );
-};
-
-export default Loading;
